refactor: user login with post action
