+++ src/api/users.js
@@ -5,11 +5,6 @@ export const signup = (name, email, password) =>
 
 const goHome = () => (location.href = '/');
 
-export const login = async (email, password) => {
-	await fetchJson('/api/users/login', { email, password });
-	goHome();
-};
-
 export const logout = async () => {
 	await fetchJson('/api/users/logout');
 	goHome();
